% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoVisualizer.R
\name{loadFASTA}
\alias{loadFASTA}
\title{Loads FASTA files (helper function)}
\usage{
loadFASTA(fastaPath)
}
\arguments{
\item{fastaPath}{Pathway of FASTA file with DNA Sequence}
}
\value{
Returns loaded FASTA in optimal format for analysis
}
\description{
Loads FASTA files (helper function)
}
\references{
Muller, K., et al. In-Line Documentation for R
R Package Roxygen2 Version 7.2.2. Comprehensive R Archive Network (CRAN),
22 July. 2022, \href{https://cran.r-project.org/web/packages/roxygen2/index.html}{Link}.
}
