% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoVisualizer-package.R
\docType{package}
\name{orthoVisualizer-package}
\alias{orthoVisualizer}
\alias{orthoVisualizer-package}
\title{orthoVisualizer: Run through DNA Sequences and Identify/Annotate Subsequence Ortholog of Interest}
\description{
Look for orthologue subsequence in multi-FASTA file of DNA, where subsequence (for specific motif/gene) is inputted by user. Annotate sequence with orthologue subsequence. Visualize abundance of such orthologous gene/motif (ex. barplot).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/karynkomatsu/orthoVisualizer}
  \item Report bugs at \url{https://github.com/karynkomatsu/orthoVisualizer/issues}
}

}
\keyword{internal}
