% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeSeq.R
\name{quantSeq}
\alias{quantSeq}
\title{Count Ortholog (gene/motif) Occurrence}
\usage{
quantSeq(fastaPath = "./inst/extdata/assembly.fasta", pattern = "CGCG")
}
\arguments{
\item{fastaPath}{is the filepath for input FASTA file containing sequences}

\item{pattern}{is the ortholog gene/motif subsequence to look for}
}
\value{
tibble of "assigned id", "header", "number of matches",
"ratio of matches"... of all sequences in FASTA file
}
\description{
Count Ortholog (gene/motif) Occurrence
}
