% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoVisualizer.R
\name{annotateSeq}
\alias{annotateSeq}
\title{Annotate Ortholog Subsequence (gene/motif) Occurrence}
\usage{
annotateSeq(fastaPath, pattern)
}
\arguments{
\item{fastaPath}{is a character string indicating the file path for input
FASTA file containing DNA sequences.}

\item{pattern}{is a character string indicating the ortholog gene/motif
subsequence to search for .}
}
\value{
Returns invisible NULL. Prints each sequence in the FASTA file with
annotation.
}
\description{
A function that prints each sequence in FASTA file and annotates every
occurrence of user-inputted ortholog pattern subsequence.
}
\examples{

#Example 1:
#Print every sequence in FASTA file containing 1 DNA sequence and annotate
#every occurrences of target orthologous binding motif pattern "CGCG".
# Get raw data
placenta_path <-  system.file("extdata", "P_Placenta_seq.fasta", package = "orthoVisualizer")
annotateSeq(fastaPath = placenta_path, pattern = "CGCG")

#Example 2:
#Print every sequence in FASTA file containing 6 DNA sequences and annotate
#every occurrences of target orthologous binding motif pattern "CGCG".
# Get raw data
DNA_seq_path <-  system.file("extdata", "DNA_seq.fasta", package = "orthoVisualizer")
annotateSeq(fastaPath = DNA_seq_path, pattern = "CGCG")

}
