% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeSeq.R
\name{annotateSeq}
\alias{annotateSeq}
\title{Annotate Ortholog (gene/motif) Occurrence}
\usage{
annotateSeq(fastaPath = "./inst/extdata/assembly.fasta", pattern = "CGCG")
}
\arguments{
\item{fastaPath}{is the filepath for input FASTA file containing sequences}

\item{pattern}{is the ortholog gene/motif subsequence to look for}
}
\value{
invisible NULL
}
\description{
Annotate Ortholog (gene/motif) Occurrence
}
