% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeSeq.R
\name{freqRatioSeq}
\alias{freqRatioSeq}
\title{Plot Ortholog (gene/motif) Occurrence Ratio}
\usage{
freqRatioSeq(fastaPath = "./inst/extdata/assembly.fasta", pattern = "CGCG")
}
\arguments{
\item{fastaPath}{is the filepath for input FASTA file containing sequences}

\item{pattern}{is the ortholog gene/motif subsequence to look for}
}
\value{
barplot of Frequency Ratio for all sequences in FASTA file
}
\description{
Plot Ortholog (gene/motif) Occurrence Ratio
}
