% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeSeq.R
\name{loadFASTA}
\alias{loadFASTA}
\title{Loads FASTA files (helper function)}
\usage{
loadFASTA(fastaPath)
}
\arguments{
\item{fastaPath}{Pathway of FASTA file with DNA Sequence}
}
\value{
Returns loaded FASTA in optimal format for analysis
}
\description{
Loads FASTA files (helper function)
}
\references{
Morgan, M. Access the Bioconductor Project Package Repository
R Package BiocManager Version 1.30.19. Comprehensive R Archive Network (CRAN),
25 Oct. 2022, \href{https://cran.r-project.org/web/packages/BiocManager/index.html}{Link}.

Huber, W. et al. (2015). Orchestrating high-throughput genomic analysis with
Bioconductor. Nature Methods, 12(2), 115–121.
\href{http://www.nature.com/nmeth/journal/v12/n2/full/nmeth.3252.html}{Link}.
}
