% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoVisualizer.R
\name{printAnnotate}
\alias{printAnnotate}
\title{Print Annotation of Ortholog Subsequence (gene/motif) Occurrence}
\usage{
printAnnotate(header, seq, pattern, motifIndex)
}
\arguments{
\item{header}{is the header of sequence}

\item{seq}{is the DNA sequence to print annotation}

\item{pattern}{is a character string indicating the ortholog gene/motif
subsequence to search for .}

\item{motifIndex}{is list of indices of where the pattern is found in seq}
}
\value{
Returns invisible NULL. Prints each sequence in the FASTA file with
annotation.
}
\description{
Print Annotation of Ortholog Subsequence (gene/motif) Occurrence
}
